% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subst.R
\name{doSubs}
\alias{doSubs}
\title{transform fixed template into program using user parameter settings}
\usage{
doSubs(tsvname, clustVblName = "Clusters", numtrees = 100,
  numthread = 1, num_inf_genes = 5, num_top_ranked = 3)
}
\arguments{
\item{tsvname}{character(1)}
}
\description{
transform fixed template into program using user parameter settings
}
